clustering IT

+++ src/main/java/com/b5m/msgpack/ClusterInfo.java
@@ -6,7 +6,7 @@ import org.msgpack.annotation.Message;
 @Message
 public class ClusterInfo {
 	public String clustername;
-	public Long clusterDocNum;
-	public Long clusterHash;
+	public int clusterDocNum;
+	public int clusterHash;
 	public Map<Integer, Float> pows;
 }
